requirements.txt
openai
pinecone-client
langchain
streamlit
arize
pandas
python-dotenv
